+++ lib/service/Config.dart
@@ -92,6 +92,7 @@ class Config extends rokwire.Config {
   String? get illiniCashTosUrl       => JsonUtils.stringValue(otherUniversityServices['illini_cash_tos_url']);
   String? get myIlliniUrl            => JsonUtils.stringValue(otherUniversityServices['myillini_url']);
   String? get feedbackUrl            => JsonUtils.stringValue(otherUniversityServices['feedback_url']);
+  String? get studentSelfServiceUrl  => JsonUtils.stringValue(otherUniversityServices['student_self_service_url']);
   String? get crisisHelpUrl          => JsonUtils.stringValue(otherUniversityServices['crisis_help_url']);
   String? get privacyPolicyUrl       => JsonUtils.stringValue(otherUniversityServices['privacy_policy_url']);
   String? get padaapiUrl             => JsonUtils.stringValue(otherUniversityServices['padaapi_url']);
@@ -160,6 +161,9 @@ class Config extends rokwire.Config {
   int  get homeCampusHighlightsCount => JsonUtils.intValue(settings['homeCampusHighlightsCount']) ?? 3;
   int  get homeCampusRemindersCount  => JsonUtils.intValue(settings['homeCampusRemindersCount']) ?? 3;
   int  get homeFavoriteItemsCount    => JsonUtils.intValue(settings['homeFavoriteItemsCount']) ?? 3;
+  int  get homeAthleticsNewsCount    => JsonUtils.intValue(settings['homeAthleticsNewsCount']) ?? 3;
+  int  get homeAthleticsEventsCount  => JsonUtils.intValue(settings['homeAthleticsEventsCount']) ?? 3;
+  int  get homeAthleticsTeamsCount   => JsonUtils.intValue(settings['homeAthleticsTeamsCount']) ?? 2;
   int  get homeRecentItemsCount      => JsonUtils.intValue(settings['homeRecentItemsCount']) ?? 3;
   int  get homeUpcomingEventsCount   => JsonUtils.intValue(settings['homeUpcomingEventsCount']) ?? 5;
   int  get recentItemsCount          => JsonUtils.intValue(settings['recentItemsCount']) ?? 32;